Don't hard-code "/cache-enabler" in advanced-cache.php
When the `advanced-cache.php` file is loaded, it was previously looking for `WP_PLUGIN_DIR . '/cache-enabler'`, which would prevent page caching from working if the plugin was not installed here for whatever reason.

This commit enables the path to be adjusted by defining a `CACHE_ENABLER_PLUGIN_DIR` constant in a site's `wp-config.php` file.